test: split oversized test files to ≤500 lines each
- test_tools.py -> + test_tools_crud.py (shared fixtures in helpers/tool_fixtures.py)
- test_integration_e2e.py -> test_e2e_lifecycle.py + test_e2e_protocol.py
- test_data_formatting.py -> test_record_formatting.py + test_dataset_formatting.py
- test_server_foundation.py -> + test_server_runtime.py
- test_performance.py -> + test_performance_cache.py
- test_json2_connection.py -> + test_json2_orm.py (helpers/json2_fixtures.py)
- test_error_handling.py -> + test_error_logging.py

Classes moved verbatim; collected test count unchanged (480 unit).
