+++ tests/test_server_runtime.py
@@ -1,0 +1,245 @@
+"""Tests for server runtime: integration, main entry point, and FastMCP app.
+
+Split out of tests/test_server_foundation.py to keep file sizes manageable.
+"""
+
+import asyncio
+import os
+from pathlib import Path
+from unittest.mock import Mock, patch
+
+import pytest
+
+from mcp_server_odoo.config import OdooConfig
+from mcp_server_odoo.odoo_connection import OdooConnectionError
+from mcp_server_odoo.server import SERVER_VERSION, OdooMCPServer
+
+
+class TestServerIntegration:
+    """Integration tests with real .env configuration."""
+
+    @pytest.mark.integration
+    def test_server_with_env_file(self, tmp_path, monkeypatch):
+        """Test server initialization with .env file in isolated environment."""
+        # Import modules we need
+        from mcp_server_odoo.config import load_config, reset_config
+
+        # Store original working directory
+        original_cwd = os.getcwd()
+
+        # Create a test .env file in tmp directory
+        env_file = tmp_path / ".env"
+        env_file.write_text("""
+ODOO_URL=http://localhost:8069
+ODOO_API_KEY=test_integration_key
+ODOO_DB=test_integration_db
+ODOO_MCP_LOG_LEVEL=DEBUG
+""")
+
+        try:
+            # Change to temp directory to isolate from project .env
+            os.chdir(tmp_path)
+
+            # Clear all environment variables that might interfere
+            for key in [
+                "ODOO_URL",
+                "ODOO_API_KEY",
+                "ODOO_DB",
+                "ODOO_MCP_LOG_LEVEL",
+                "ODOO_USER",
+                "ODOO_PASSWORD",
+            ]:
+                monkeypatch.delenv(key, raising=False)
+
+            # Reset config singleton
+            reset_config()
+
+            # Load config explicitly from our test .env file
+            # This ensures we're loading from the tmp directory's .env
+            config = load_config(env_file)
+
+            # Create server with the loaded config
+            server = OdooMCPServer(config)
+
+            assert server.config.url == "http://localhost:8069"
+            assert server.config.api_key == "test_integration_key"
+            assert server.config.database == "test_integration_db"
+            assert server.config.log_level == "DEBUG"
+
+        finally:
+            os.chdir(original_cwd)
+            reset_config()  # Reset again for other tests
+
+    @pytest.mark.integration
+    @pytest.mark.asyncio
+    async def test_real_odoo_connection(self):
+        """Test with real Odoo connection using .env credentials.
+
+        This test requires a running Odoo server with valid credentials
+        in the .env file.
+        """
+        # Skip if no .env file exists
+        if not Path(".env").exists():
+            pytest.skip("No .env file found for integration test")
+
+        # Import and reset config to ensure clean state
+        from mcp_server_odoo.config import reset_config
+
+        reset_config()
+
+        # Load environment
+        from dotenv import load_dotenv
+
+        load_dotenv()
+
+        # Check if required env vars are set
+        if not os.getenv("ODOO_URL"):
+            pytest.skip("ODOO_URL not set in environment")
+
+        server = None
+        try:
+            # Create server with real config
+            server = OdooMCPServer()
+
+            # Test connection
+            server._ensure_connection()
+
+            # If we get here, connection was successful
+            assert server.connection is not None
+
+            # Clean up
+            server._cleanup_connection()
+
+        except OdooConnectionError as e:
+            # Connection errors are expected if Odoo is not running
+            pytest.skip(f"Integration test skipped (Odoo not available): {e}")
+        except Exception as e:
+            # Other exceptions might indicate a test issue
+            import traceback
+
+            pytest.skip(
+                f"Integration test skipped (unexpected error): {type(e).__name__}: {e}\n{traceback.format_exc()}"
+            )
+        finally:
+            # Always reset config for other tests
+            reset_config()
+
+
+class TestMainEntry:
+    """Test the __main__ entry point."""
+
+    def test_help_flag(self, capsys):
+        """Test --help flag."""
+        from mcp_server_odoo.__main__ import main
+
+        # argparse raises SystemExit for --help
+        try:
+            exit_code = main(["--help"])
+            assert exit_code == 0
+        except SystemExit as e:
+            assert e.code == 0
+
+        captured = capsys.readouterr()
+        # Help output goes to stdout by default from argparse
+        help_output = captured.out or captured.err
+        assert "Odoo MCP Server" in help_output
+        assert "ODOO_URL" in help_output
+
+    def test_version_flag(self, capsys):
+        """Test --version flag."""
+        from mcp_server_odoo.__main__ import main
+
+        # argparse raises SystemExit for --version
+        try:
+            exit_code = main(["--version"])
+            assert exit_code == 0
+        except SystemExit as e:
+            assert e.code == 0
+
+        captured = capsys.readouterr()
+        # Version output goes to stdout by default from argparse
+        version_output = captured.out or captured.err
+        assert f"odoo-mcp-server v{SERVER_VERSION}" in version_output
+
+    def test_main_with_invalid_config(self, capsys, monkeypatch):
+        """Test main with invalid configuration."""
+        from mcp_server_odoo.__main__ import main
+
+        # Set invalid config
+        monkeypatch.setenv("ODOO_URL", "")  # Empty URL
+
+        exit_code = main([])
+
+        assert exit_code == 1
+
+        captured = capsys.readouterr()
+        assert "Configuration error" in captured.err
+
+    def test_main_with_valid_config(self, monkeypatch):
+        """Test main with valid configuration."""
+        from mcp_server_odoo.__main__ import main
+
+        # Set valid config
+        monkeypatch.setenv("ODOO_URL", "http://localhost:8069")
+        monkeypatch.setenv("ODOO_API_KEY", "test_key")
+
+        # Mock the server and its run_stdio method
+        with patch("mcp_server_odoo.__main__.OdooMCPServer") as mock_server_class:
+            mock_server = Mock()
+
+            # Create a coroutine that completes immediately
+            async def mock_run_stdio():
+                pass
+
+            mock_server.run_stdio = mock_run_stdio
+            mock_server_class.return_value = mock_server
+
+            # Mock asyncio.run to execute synchronously
+            def mock_asyncio_run(coro):
+                # Run the coroutine to completion
+                loop = asyncio.new_event_loop()
+                try:
+                    return loop.run_until_complete(coro)
+                finally:
+                    loop.close()
+
+            with patch("asyncio.run", side_effect=mock_asyncio_run):
+                exit_code = main([])
+
+                assert exit_code == 0
+                mock_server_class.assert_called_once()
+
+
+class TestFastMCPApp:
+    """Test the FastMCP app configuration."""
+
+    @pytest.fixture
+    def valid_config(self):
+        """Create a valid test configuration."""
+        return OdooConfig(
+            url=os.getenv("ODOO_URL", "http://localhost:8069"),
+            api_key="test_api_key_12345",
+            database="test_db",
+            log_level="INFO",
+            default_limit=10,
+            max_limit=100,
+        )
+
+    def test_fastmcp_app_creation(self, valid_config):
+        """Test that FastMCP app is properly created."""
+        server = OdooMCPServer(valid_config)
+
+        assert server.app is not None
+        assert server.app.name == "odoo-mcp-server"
+        assert "Odoo" in server.app.instructions
+        assert len(server.app.instructions) > 100
+
+    def test_fastmcp_app_has_required_methods(self, valid_config):
+        """Test that FastMCP app has required methods."""
+        server = OdooMCPServer(valid_config)
+
+        # Check that required methods exist
+        assert hasattr(server.app, "run_stdio_async")
+        assert hasattr(server.app, "resource")
+        assert hasattr(server.app, "tool")
+        assert hasattr(server.app, "prompt")
