+++ tests/test_tools.py
@@ -1,68 +1,38 @@
-"""Test suite for MCP tools functionality."""
+"""Test suite for MCP tools functionality.
+
+Access-control and error-path tests for the tool handler live in
+tests/test_tools_crud.py. Shared fixtures live in tests/helpers/tool_fixtures.py.
+"""
 
 from unittest.mock import MagicMock
 
 import pytest
 from mcp.server.fastmcp import FastMCP
 
-from mcp_server_odoo.access_control import AccessControlError, AccessController
+from mcp_server_odoo.access_control import AccessController
 from mcp_server_odoo.config import OdooConfig
-from mcp_server_odoo.error_handling import (
-    ValidationError,
-)
-from mcp_server_odoo.odoo_connection import OdooConnection, OdooConnectionError
+from mcp_server_odoo.odoo_connection import OdooConnection
 from mcp_server_odoo.tools import OdooToolHandler, register_tools
+from tests.helpers.tool_fixtures import (
+    handler,
+    mock_access_controller,
+    mock_app,
+    mock_connection,
+    valid_config,
+)
+
+# Re-export shared fixtures so pytest can resolve them in this module
+__all__ = [
+    "handler",
+    "mock_access_controller",
+    "mock_app",
+    "mock_connection",
+    "valid_config",
+]
 
 
 class TestOdooToolHandler:
     """Test cases for OdooToolHandler class."""
-
-    @pytest.fixture
-    def mock_app(self):
-        """Create a mock FastMCP app."""
-        app = MagicMock(spec=FastMCP)
-        # Store registered tools
-        app._tools = {}
-
-        def tool_decorator(**kwargs):
-            def decorator(func):
-                # Store the function in our tools dict
-                app._tools[func.__name__] = func
-                return func
-
-            return decorator
-
-        app.tool = tool_decorator
-        return app
-
-    @pytest.fixture
-    def mock_connection(self):
-        """Create a mock OdooConnection."""
-        connection = MagicMock(spec=OdooConnection)
-        connection.is_authenticated = True
-        return connection
-
-    @pytest.fixture
-    def mock_access_controller(self):
-        """Create a mock AccessController."""
-        controller = MagicMock(spec=AccessController)
-        return controller
-
-    @pytest.fixture
-    def valid_config(self):
-        """Create a valid config."""
-        return OdooConfig(
-            url="http://localhost:8069",
-            api_key="test_api_key",
-            database="test_db",
-            default_limit=100,
-            max_limit=500,
-        )
-
-    @pytest.fixture
-    def handler(self, mock_app, mock_connection, mock_access_controller, valid_config):
-        """Create an OdooToolHandler instance."""
-        return OdooToolHandler(mock_app, mock_connection, mock_access_controller, valid_config)
 
     def test_handler_initialization(self, handler, mock_app):
         """Test handler is properly initialized."""
@@ -121,59 +91,6 @@ class TestOdooToolHandler:
         mock_connection.search.assert_called_once_with(
             "res.partner", [["is_company", "=", True]], limit=3, offset=0, order="name asc"
         )
-
-    @pytest.mark.asyncio
-    async def test_search_records_access_denied(
-        self, handler, mock_connection, mock_access_controller, mock_app
-    ):
-        """Test search_records with access denied."""
-        # Setup mocks
-        mock_access_controller.validate_model_access.side_effect = AccessControlError(
-            "Access denied"
-        )
-
-        # Get the registered search_records function
-        search_records = mock_app._tools["search_records"]
-
-        # Call the tool and expect error
-        with pytest.raises(ValidationError) as exc_info:
-            await search_records(model="res.partner", domain=[], fields=None, limit=10)
-
-        assert "Access denied" in str(exc_info.value)
-
-    @pytest.mark.asyncio
-    async def test_search_records_not_authenticated(
-        self, handler, mock_connection, mock_access_controller, mock_app
-    ):
-        """Test search_records when not authenticated."""
-        # Setup mocks
-        mock_connection.is_authenticated = False
-
-        # Get the registered search_records function
-        search_records = mock_app._tools["search_records"]
-
-        # Call the tool and expect error
-        with pytest.raises(ValidationError) as exc_info:
-            await search_records(model="res.partner")
-
-        assert "Not authenticated" in str(exc_info.value)
-
-    @pytest.mark.asyncio
-    async def test_search_records_connection_error(
-        self, handler, mock_connection, mock_access_controller, mock_app
-    ):
-        """Test search_records with connection error."""
-        # Setup mocks
-        mock_connection.search_count.side_effect = OdooConnectionError("Connection lost")
-
-        # Get the registered search_records function
-        search_records = mock_app._tools["search_records"]
-
-        # Call the tool and expect error
-        with pytest.raises(ValidationError) as exc_info:
-            await search_records(model="res.partner")
-
-        assert "Connection error" in str(exc_info.value)
 
     @pytest.mark.asyncio
     async def test_search_records_with_domain_operators(
@@ -282,26 +199,6 @@ class TestOdooToolHandler:
         mock_connection.search_count.assert_called_with("res.partner", expected_domain)
 
     @pytest.mark.asyncio
-    async def test_search_records_with_invalid_json_domain(
-        self, handler, mock_connection, mock_access_controller, mock_app
-    ):
-        """Test search_records with invalid JSON string domain."""
-        # Setup mocks
-        mock_access_controller.validate_model_access.return_value = None
-
-        # Get the registered search_records function
-        search_records = mock_app._tools["search_records"]
-
-        # Invalid JSON string
-        invalid_domain = '[["is_company", "=", true'  # Missing closing brackets
-
-        # Should raise ValidationError
-        with pytest.raises(ValidationError) as exc_info:
-            await search_records(model="res.partner", domain=invalid_domain, limit=5)
-
-        assert "Invalid search criteria format" in str(exc_info.value)
-
-    @pytest.mark.asyncio
     async def test_search_records_with_string_fields(
         self, handler, mock_connection, mock_access_controller, mock_app
     ):
@@ -392,76 +289,6 @@ class TestOdooToolHandler:
         mock_connection.read.assert_called_once_with("res.partner", [123], ["name", "email"])
 
     @pytest.mark.asyncio
-    async def test_get_record_not_found(
-        self, handler, mock_connection, mock_access_controller, mock_app
-    ):
-        """Test get_record when record doesn't exist."""
-        # Setup mocks
-        mock_connection.read.return_value = []
-
-        # Get the registered get_record function
-        get_record = mock_app._tools["get_record"]
-
-        # Call the tool and expect error
-        with pytest.raises(ValidationError) as exc_info:
-            await get_record(model="res.partner", record_id=999)
-
-        assert "Record not found" in str(exc_info.value)
-
-    @pytest.mark.asyncio
-    async def test_get_record_access_denied(
-        self, handler, mock_connection, mock_access_controller, mock_app
-    ):
-        """Test get_record with access denied."""
-        # Setup mocks
-        mock_access_controller.validate_model_access.side_effect = AccessControlError(
-            "Access denied"
-        )
-
-        # Get the registered get_record function
-        get_record = mock_app._tools["get_record"]
-
-        # Call the tool and expect error
-        with pytest.raises(ValidationError) as exc_info:
-            await get_record(model="res.partner", record_id=1)
-
-        assert "Access denied" in str(exc_info.value)
-
-    @pytest.mark.asyncio
-    async def test_get_record_not_authenticated(
-        self, handler, mock_connection, mock_access_controller, mock_app
-    ):
-        """Test get_record when not authenticated."""
-        # Setup mocks
-        mock_connection.is_authenticated = False
-
-        # Get the registered get_record function
-        get_record = mock_app._tools["get_record"]
-
-        # Call the tool and expect error
-        with pytest.raises(ValidationError) as exc_info:
-            await get_record(model="res.partner", record_id=1)
-
-        assert "Not authenticated" in str(exc_info.value)
-
-    @pytest.mark.asyncio
-    async def test_get_record_connection_error(
-        self, handler, mock_connection, mock_access_controller, mock_app
-    ):
-        """Test get_record with connection error."""
-        # Setup mocks
-        mock_connection.read.side_effect = OdooConnectionError("Connection lost")
-
-        # Get the registered get_record function
-        get_record = mock_app._tools["get_record"]
-
-        # Call the tool and expect error
-        with pytest.raises(ValidationError) as exc_info:
-            await get_record(model="res.partner", record_id=1)
-
-        assert "Connection error" in str(exc_info.value)
-
-    @pytest.mark.asyncio
     async def test_list_models_success(
         self, handler, mock_connection, mock_access_controller, mock_app
     ):
@@ -525,71 +352,6 @@ class TestOdooToolHandler:
 
         # Verify calls
         mock_access_controller.get_enabled_models.assert_called_once()
-
-    @pytest.mark.asyncio
-    async def test_list_models_with_permission_failures(
-        self, handler, mock_connection, mock_access_controller, mock_app
-    ):
-        """Test list_models when some models fail to get permissions."""
-        # Setup mocks for get_enabled_models
-        mock_access_controller.get_enabled_models.return_value = [
-            {"model": "res.partner", "name": "Contact"},
-            {"model": "unknown.model", "name": "Unknown Model"},
-        ]
-
-        # Setup mocks for get_model_permissions
-        from mcp_server_odoo.access_control import AccessControlError, ModelPermissions
-
-        partner_perms = ModelPermissions(
-            model="res.partner",
-            enabled=True,
-            can_read=True,
-            can_write=True,
-            can_create=False,
-            can_unlink=False,
-        )
-
-        # Configure side_effect to fail for unknown model
-        def get_perms(model):
-            if model == "res.partner":
-                return partner_perms
-            else:
-                raise AccessControlError(f"Model {model} not found")
-
-        mock_access_controller.get_model_permissions.side_effect = get_perms
-
-        # Get the registered list_models function
-        list_models = mock_app._tools["list_models"]
-
-        # Call the tool - should not fail even if some models can't get permissions
-        result = await list_models()
-
-        # Verify result structure (ModelsResult is a Pydantic model)
-        assert len(result.models) == 2
-
-        # Verify models are returned without per-model permission checks
-        partner = result.models[0]
-        assert partner.model == "res.partner"
-
-        unknown = result.models[1]
-        assert unknown.model == "unknown.model"
-
-    @pytest.mark.asyncio
-    async def test_list_models_error(
-        self, handler, mock_connection, mock_access_controller, mock_app
-    ):
-        """Test list_models with error."""
-        # Setup mocks
-        mock_access_controller.get_enabled_models.side_effect = Exception("API error")
-
-        # Get the registered list_models function
-        list_models = mock_app._tools["list_models"]
-
-        # Call the tool and expect error
-        with pytest.raises(ValidationError) as exc_info:
-            await list_models()
-
-        assert "Failed to list models" in str(exc_info.value)
 
     @pytest.mark.asyncio
     async def test_search_records_with_defaults(
